feat: re add, list tree

+++ backend/sswm/src/main/java/com/ground/sswm/usertree/dto/UserTreeDto.java
@@ -10,6 +10,6 @@ import lombok.ToString;
 public class UserTreeDto {
 
     private Long userId;
-    private int treeId;
+    private Long treeId;
     private int exp;
 }
